fast-api-app/requirements.txt
fastapi[standard]==0.134.0
uvicorn==0.41.0
sqlalchemy[asyncio]==2.0.47
asyncpg==0.31.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
